+++ src/directives/base-map-directive.ts
@@ -1,18 +1,19 @@
-import { EventEmitter, SimpleChanges, ReflectiveInjector, OnInit, OnChanges, OnDestroy } from '@angular/core';
+import { EventEmitter, SimpleChanges, OnInit, OnChanges, OnDestroy, Output } from '@angular/core';
 
 import { OptionBuilder } from '../services/option-builder';
 import { Ng2Map } from '../services/ng2-map';
-import { GeoCoder } from '../services/geo-coder';
 import { Ng2MapComponent } from '../components/ng2-map.component';
 
 export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
+  @Output() public initialized$: EventEmitter<any> = new EventEmitter();
+
   public mapObject: any; // e.g. google.maps.Marker
   public objectOptions: any; // e.g. google.maps.MarkerOptions
 
   public ng2Map: Ng2Map;
   public optionBuilder: OptionBuilder;
-  public initialized$: EventEmitter<any> = new EventEmitter();
   public libraryName: string;
+  protected _subscriptions = [];
 
   constructor(
     protected ng2MapComponent: Ng2MapComponent,
@@ -22,6 +23,7 @@ export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
   ) {
     this.ng2Map = this.ng2MapComponent['ng2Map'];
     this.optionBuilder = this.ng2MapComponent['optionBuilder'];
+    // all outputs must be initialized
     this.outputs.forEach(output => this[output] = new EventEmitter());
     this.mapObjectName = mapObjectName;
   }
@@ -57,6 +59,7 @@ export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
     // set google events listeners and emits to this outputs listeners
     this.ng2Map.setObjectEvents(this.outputs, this, 'mapObject');
 
+    this.ng2MapComponent.addToMapObjectGroup(this.mapObjectName, this.mapObject);
     this.initialized$.emit(this.mapObject);
   }
 
@@ -69,9 +72,12 @@ export abstract class BaseMapDirective implements OnInit, OnChanges, OnDestroy {
 
   // When destroyed, remove event listener, and delete this object to prevent memory leak
   ngOnDestroy() {
+    this._subscriptions.map(subscription => subscription.unsubscribe());
+    this.ng2MapComponent.removeFromMapObjectGroup(this.mapObjectName, this.mapObject);
+
     if (this.mapObject) {
       this.outputs.forEach(output => google.maps.event.clearListeners(this.mapObject, output));
-      delete this.mapObject['setMap'](null);
+      this.mapObject['setMap'](null);
       delete this.mapObject;
     }
   }